C Libraries
Custom Implementation's of Libraries created by myself for ease of use and Implementation into Coffee OS. This was also mainly for learning purposes, to see how each feature and/or commonly used coding methods/functions are built and used.

Doing this everyday is giving me a deeper insight into low-level aspects of the machine. Very cool!